separate modal and radio in react components, fixed yarn build
