+++ src/components/List/SortBySelect.tsx
@@ -1,31 +1,11 @@
 import React, { useState } from "react";
 import { Button, Modal, Radio, Space } from "antd";
 import { SortAscendingOutlined } from "@ant-design/icons";
-interface ISortByButton {
-    sortBy: string
-    refreshListItems: (sortBy: string) => void
-}
-const SortByButton = ({sortBy, refreshListItems}: ISortByButton) => {
+import OrderModal from '../Modal/OrderModal'
+
+const SortByButton = ({sortBy, refreshListItems}) => {
     const [visible, setVisible] = useState(false)
     const [ value, setValue ] = useState('asc')
-
-    const antModalHeader = {
-        borderRadius: '20px 20px 0 0',
-        background: 'green'
-    }
-
-    const handleOk = () => {
-        refreshListItems(value)
-        setVisible(!visible)
-    };
-
-    const handleCancel = () => {
-        setVisible(false);
-    };
-
-    const onChangeRadio = e => {
-        setValue(e.target.value);
-    };
 
     return (
         <>
@@ -60,120 +40,18 @@ const SortByButton = ({sortBy, refreshListItems}: ISortByButton) => {
                 </span>
             </Button>
 
-            <Modal
+            <OrderModal
                 visible={visible}
-                footer={false}
-            >
-                <p
-                    style={{
-                    color: '#111111',
-                    fontWeight: 700,
-                    fontSize: '18px',
-                    textAlign: 'center',
-                    
-                    }}
-                >
-                    Sort speeches by
-                </p>
-                <Radio.Group
-                    onChange={onChangeRadio}
-                    value={value}
-                    style={{
-                        fontSize: '30px',
-                    }}
-                >
-                    <Space direction="vertical">
-                        <Radio 
-                            value='recent'
-                        >
-                            <span style={{marginLeft:10}}>
-                                Recently added
-                            </span>
-                        </Radio>
-                        <Radio
-                            value='asc'
-                        >
-                            <span style={{marginLeft:10}}>
-                                Newest
-                            </span>
-                        </Radio>
-                        <Radio
-                            value='desc'
-                        >
-                            <span style={{marginLeft:10}}>
-                                Oldest
-                            </span>
-                        </Radio>
-                        <Radio
-                            value='most'
-                        >
-                            <span style={{marginLeft:10}}>
-                                Most reliable
-                            </span>
-                        </Radio>
-                        <Radio
-                            value='less'
-                        >
-                            <span style={{marginLeft:10}}>
-                                Less reliable
-                            </span>
-                        </Radio>
-                    </Space>
-                </Radio.Group>
-                <div style={{
-                    position: 'absolute',
-                    left: 30,
-                    bottom: 24,
-                    display: 'flex',
-                    justifyContent: 'space-between'
-                }}>
-                    <Button
-                        type="text"
-                        onClick={handleCancel}
-                        style={{
-                            display: 'flex',
-                            alignItems: 'center',
-                            height: 40,
-                            width: 120,
-                            justifyContent: 'center'
-                        }}
-                    >
-                        <span
-                            style={{
-                                color: '#2D77A3',
-                                textDecorationLine: 'underline',
-                                textAlign: 'center',
-                                fontWeight: 700,
-                                fontSize: 14,
-                            }}
-                        >
-                            Cancel
-                        </span>
-                    </Button>
-
-                    <Button
-                        onClick={handleOk}
-                        shape='round'
-                        style={{
-                            background: '#2D77A3',
-                            width: 120,
-                            height: 40,
-                        }}
-                    >
-                        <span
-                            style={{
-                                color: '#fff',
-                                textAlign: 'center',
-                                fontWeight: 700,
-                                fontSize: 14,
-                            }}
-                        >
-                            Done
-                        </span>
-                    </Button>
-                </div>
-            </Modal>
+                value={value}
+                setValue={setValue}
+                handleOk={() => {
+                    refreshListItems(value)
+                    setVisible(!visible)
+                }}
+                handleCancel={() => setVisible(false)}
+            />
         </>
     )
 }
+
 export default SortByButton
